+++ prepareimport/prepareimport.go
@@ -1,141 +1,22 @@
 package prepareimport
 
 import (
-	"encoding/json"
-	"errors"
 	"io/ioutil"
 	"path"
-	"path/filepath"
-	"regexp"
-	"strings"
-	"time"
 )
 
-type batchKeyType string
-
-func (b batchKeyType) String() string {
-	return string(b)
-}
-
-func (b batchKeyType) Path() string {
-	return "/" + string(b) + "/"
-}
-
-// BatchKey represents a valid batch key.
-type BatchKey interface {
-	String() string
-	Path() string
-}
-
-// NewBatchKey constructs a valid batch key.
-func NewBatchKey(key string) (BatchKey, error) {
-	var isValidBatchKey = regexp.MustCompile(`^[0-9]{4}`)
-	if !isValidBatchKey.MatchString(key) {
-		return batchKeyType(""), errors.New("la clé du batch doit respecter le format requis AAMM")
+// PrepareImport generates an Admin object from files found at given pathname of the file system.
+func PrepareImport(pathname string, batchKey BatchKey, dateFinEffectif DateFinEffectif) (AdminObject, error) {
+	batchPath := path.Join(pathname, batchKey.String())
+	filenames, err := ReadFilenames(batchPath)
+	if err != nil {
+		return nil, err
 	}
-	return batchKeyType(key), nil
-}
-
-type dateFinEffectifType struct {
-	time.Time
-}
-
-// MongoDate converts a dateFinEffectif to a MongoDate value.
-func (dateFinEffectif *dateFinEffectifType) MongoDate() MongoDate {
-	return MongoDate{strings.Replace(dateFinEffectif.Format(time.RFC3339), "Z", ".000+0000", 1)}
-}
-
-// NewDateFinEffectif creates a valid dateFinEffectif.
-func NewDateFinEffectif(date time.Time) dateFinEffectifType {
-	// TODO: validate the value
-	return dateFinEffectifType{date}
-}
-
-// AdminObject represents a document going to be stored in the Admin db collection.
-type AdminObject map[string]interface{}
-
-// IDProperty represents the "_id" property of an Admin object.
-type IDProperty struct {
-	Key  BatchKey `json:"key"`
-	Type string   `json:"type"`
-}
-
-// FilesProperty represents the "files" property of an Admin object.
-type FilesProperty map[ValidFileType][]string
-
-// MongoDate represents a date that can be serialized for MongoDB.
-type MongoDate struct {
-	date string
-}
-
-// MarshalJSON will be called when serializing a date for MongoDB.
-func (mongoDate MongoDate) MarshalJSON() ([]byte, error) {
-	return json.Marshal(map[string]string{"$date": mongoDate.date})
-}
-
-// UnmarshalJSON will parse a MongoDate from a MongoDB date object.
-func (mongoDate *MongoDate) UnmarshalJSON(data []byte) error {
-	var dateObj map[string]string
-	if err := json.Unmarshal(data, &dateObj); err != nil {
-		return err
+	augmentedFiles := []DataFile{}
+	for _, file := range filenames {
+		augmentedFiles = append(augmentedFiles, AugmentDataFile(file, batchPath))
 	}
-	mongoDate.date = dateObj["$date"]
-	return nil
-}
-
-// ParamProperty represents the "param" property of an Admin object.
-type ParamProperty struct {
-	DateDebut       MongoDate `json:"date_debut"`
-	DateFin         MongoDate `json:"date_fin"`
-	DateFinEffectif MongoDate `json:"date_fin_effectif"`
-}
-
-// DataFile represents a Data File to be imported, and allows to determine its type and name.
-type DataFile interface {
-	GetFilename() string           // the name as it will be stored in Admin
-	DetectFileType() ValidFileType // returns the type of that file (e.g. DEBIT)
-}
-
-// SimpleDataFile is a DataFile which type can be determined without requiring a metadata file (e.g. well-named csv file).
-type SimpleDataFile struct {
-	filename string
-}
-
-// DetectFileType returns the type of that file (e.g. DEBIT).
-func (dataFile SimpleDataFile) DetectFileType() ValidFileType {
-	return ExtractFileTypeFromFilename(dataFile.filename)
-}
-
-// GetFilename returns the name as it will be stored in Admin.
-func (dataFile SimpleDataFile) GetFilename() string {
-	return dataFile.filename
-}
-
-// UploadedDataFile is a DataFile which type can be determined thanks to a metadata file (e.g. bin+info files).
-type UploadedDataFile struct {
-	filename string
-	path     string
-}
-
-// DetectFileType returns the type of that file (e.g. DEBIT).
-func (dataFile UploadedDataFile) DetectFileType() ValidFileType {
-	metaFilepath := filepath.Join(dataFile.path, dataFile.filename+".info")
-	fileinfo := LoadMetadata(metaFilepath)
-	return ExtractFileTypeFromMetadata(metaFilepath, fileinfo) // e.g. "Sigfaible_debits.csv"
-}
-
-// GetFilename returns the name as it will be stored in Admin.
-func (dataFile UploadedDataFile) GetFilename() string {
-	return dataFile.filename
-}
-
-// UnsupportedFilesError is an Error object that lists files that were not supported.
-type UnsupportedFilesError struct {
-	UnsupportedFiles []string
-}
-
-func (err UnsupportedFilesError) Error() string {
-	return "unsupported: " + strings.Join(err.UnsupportedFiles, ", ")
+	return PopulateAdminObject(augmentedFiles, batchKey, dateFinEffectif)
 }
 
 // ReadFilenames returns the name of files found at the provided path.
@@ -149,96 +30,4 @@ func ReadFilenames(path string) ([]string, error) {
 		files = append(files, file.Name())
 	}
 	return files, nil
-}
-
-// AugmentDataFile returns a SimpleDataFile or UploadedDataFile (if metadata had to be loaded).
-func AugmentDataFile(file string, pathname string) DataFile {
-	if !strings.Contains(file, ".") { // "bin" files have no extension => no dot in their filename
-		return UploadedDataFile{file, pathname}
-	}
-	return SimpleDataFile{file}
-}
-
-// PrepareImport generates an Admin object from files found at given pathname of the file system.
-func PrepareImport(pathname string, batchKey BatchKey, dateFinEffectif dateFinEffectifType) (AdminObject, error) {
-	batchPath := path.Join(pathname, batchKey.String())
-	filenames, err := ReadFilenames(batchPath)
-	if err != nil {
-		return nil, err
-	}
-	augmentedFiles := []DataFile{}
-	for _, file := range filenames {
-		augmentedFiles = append(augmentedFiles, AugmentDataFile(file, batchPath))
-	}
-	return PopulateAdminObject(augmentedFiles, batchKey, dateFinEffectif)
-}
-
-// PopulateAdminObject populates an AdminObject, given a list of data files.
-func PopulateAdminObject(augmentedFilenames []DataFile, batchKey BatchKey, dateFinEffectif dateFinEffectifType) (AdminObject, error) {
-
-	filesProperty, unsupportedFiles := PopulateFilesProperty(augmentedFilenames, batchKey.Path())
-	var err error
-	if len(unsupportedFiles) > 0 {
-		err = UnsupportedFilesError{unsupportedFiles}
-	}
-	var completeTypes = []ValidFileType{}
-	for _, typeName := range defaultCompleteTypes {
-		if _, ok := filesProperty[typeName]; ok {
-			completeTypes = append(completeTypes, typeName)
-		}
-	}
-	// { "date_debut" : { "$date" : "2014-01-01T00:00:00.000+0000" }, "date_fin" : { "$date" : "2018-12-01T00:00:00.000+0000" }, "date_fin_effectif" : { "$date" : "2018-06-01T00:00:00.000+0000" } }
-
-	paramProperty := ParamProperty{
-		DateDebut:       MongoDate{"2014-01-01T00:00:00.000+0000"},
-		DateFin:         MongoDate{"20" + batchKey.String()[0:2] + "-" + batchKey.String()[2:4] + "-01T00:00:00.000+0000"},
-		DateFinEffectif: dateFinEffectif.MongoDate(),
-	}
-
-	return AdminObject{
-		"_id":            IDProperty{batchKey, "batch"},
-		"files":          filesProperty,
-		"complete_types": completeTypes,
-		"param":          paramProperty,
-	}, err
-}
-
-// LoadMetadata returns the metadata of a bin file (without extension), by reading the given .info file.
-func LoadMetadata(filepath string) UploadedFileMeta {
-
-	// read file
-	data, err := ioutil.ReadFile(filepath)
-	if err != nil {
-		panic(err)
-	}
-
-	// unmarshall data from json
-	var uploadedFileMeta UploadedFileMeta
-	err = json.Unmarshal(data, &uploadedFileMeta)
-	if err != nil {
-		panic(err)
-	}
-
-	return uploadedFileMeta
-}
-
-// PopulateFilesProperty populates the "files" property of an Admin object, given a list of Data files.
-func PopulateFilesProperty(filenames []DataFile, prefix string) (FilesProperty, []string) {
-	filesProperty := FilesProperty{}
-	unsupportedFiles := []string{}
-	for _, filename := range filenames {
-		filetype := filename.DetectFileType()
-
-		if filetype == "" {
-			if !strings.HasSuffix(filename.GetFilename(), ".info") {
-				unsupportedFiles = append(unsupportedFiles, prefix+filename.GetFilename())
-			}
-			continue
-		}
-		if _, exists := filesProperty[filetype]; !exists {
-			filesProperty[filetype] = []string{}
-		}
-		filesProperty[filetype] = append(filesProperty[filetype], prefix+filename.GetFilename())
-	}
-	return filesProperty, unsupportedFiles
 }
